+++ plt/plt-scheduler/src/scheduler/plt_scheduler.rs
@@ -12,7 +12,7 @@ use crate::scheduler::TransactionFailure;
 use crate::transaction_execution::TransactionExecution;
 use concordium_base::base::AccountIndex;
 use concordium_base::common::cbor::{self};
-use concordium_base::protocol_level_locks::LockId;
+use concordium_base::protocol_level_locks::{LockId, LockRecipients as CborLockRecipients};
 use concordium_base::protocol_level_tokens::meta_operations::{
     LockOperation, MetaLockCancelDetails, MetaLockCreateDetails, MetaLockFundDetails,
     MetaLockReturnDetails, MetaLockSendDetails,
@@ -29,7 +29,7 @@ use plt_block_state::entity::block_state::p11::BlockStateP11;
 use plt_block_state::entity::{EntityContext, EntityContextTypes};
 use plt_block_state::failure::BlockStateFailure;
 use plt_block_state::persistent::protocol_level_locks::p11::{
-    LockConfiguration, LockControllerConfig,
+    LockConfiguration, LockControllerConfig, LockRecipients,
 };
 use plt_block_state::persistent::protocol_level_tokens::p9::TokenConfiguration;
 use plt_scheduler_types::types::events::{self, BlockItemEvent};
@@ -106,15 +106,15 @@ where
 
     let lock_configuration = lock.lock_configuration(context)?;
     if lock_configuration
-        .expiry()
+        .expiry
         .is_expired(transaction_execution.timestamp())
     {
         return Err(
-            TransactionRejectReason::LockExpired(lock_configuration.lock_id().clone()).into(),
+            TransactionRejectReason::LockExpired(lock_configuration.lock_id.clone()).into(),
         );
     }
 
-    lock_configuration.controller().validate_operation(
+    lock_configuration.controller.validate_operation(
         &bsq,
         transaction_execution.sender_account_address(),
         transaction_execution.sender_account(),
@@ -134,7 +134,7 @@ where
         &mut token,
         transaction_execution.sender_account(),
         transaction_execution.sender_account_address(),
-        lock_configuration.lock_id(),
+        &lock_configuration.lock_id,
         raw_amount,
         memo,
     )? {
@@ -184,11 +184,11 @@ where
 
     let lock_configuration = lock.lock_configuration(context)?;
     if lock_configuration
-        .expiry()
+        .expiry
         .is_expired(transaction_execution.timestamp())
     {
         return Err(
-            TransactionRejectReason::LockExpired(lock_configuration.lock_id().clone()).into(),
+            TransactionRejectReason::LockExpired(lock_configuration.lock_id.clone()).into(),
         );
     }
 
@@ -223,15 +223,18 @@ where
         return Err(token_update_error_reject_reason(&token_configuration, err));
     }
 
-    if !lock_configuration.is_recipient(&recipient.account_index()) {
+    if !lock_configuration
+        .recipients
+        .is_recipient(&recipient.account_index())
+    {
         return Err(TransactionRejectReason::LockRecipientNotPermitted(
-            lock_configuration.lock_id().clone(),
+            lock_configuration.lock_id.clone(),
             recipient_address,
         )
         .into());
     }
 
-    lock_configuration.controller().validate_operation(
+    lock_configuration.controller.validate_operation(
         &bsq,
         transaction_execution.sender_account_address(),
         transaction_execution.sender_account(),
@@ -249,7 +252,7 @@ where
         source_address,
         &recipient,
         recipient_address,
-        lock_configuration.lock_id(),
+        &lock_configuration.lock_id,
         raw_amount,
         memo,
     )?
@@ -298,11 +301,11 @@ where
 
     let lock_configuration = lock.lock_configuration(context)?;
     if lock_configuration
-        .expiry()
+        .expiry
         .is_expired(transaction_execution.timestamp())
     {
         return Err(
-            TransactionRejectReason::LockExpired(lock_configuration.lock_id().clone()).into(),
+            TransactionRejectReason::LockExpired(lock_configuration.lock_id.clone()).into(),
         );
     }
 
@@ -311,7 +314,7 @@ where
         .account_by_address(&source_address)
         .map_err(|_| TransactionRejectReason::InvalidAccountReference(source_address))?;
 
-    lock_configuration.controller().validate_operation(
+    lock_configuration.controller.validate_operation(
         &bsq,
         transaction_execution.sender_account_address(),
         transaction_execution.sender_account(),
@@ -331,7 +334,7 @@ where
         &mut token,
         source.account_index(),
         source_address,
-        lock_configuration.lock_id(),
+        &lock_configuration.lock_id,
         raw_amount,
         memo,
     )?
@@ -373,27 +376,40 @@ where
         context: context.clone(),
     };
 
-    let config = details.config;
     let account_index = transaction_execution.sender_account().account_index();
     let sequence_number = transaction_execution.transaction_sequence_number();
     let creation_order = transaction_execution.next_lock_creation_order();
     let lock_id = LockId::new(account_index, sequence_number, creation_order);
-    let controller = LockController::new(&bsq, config.controller)?;
+    let concordium_base::protocol_level_locks::LockConfig {
+        recipients,
+        expiry,
+        controller: controller_config,
+    } = details.config;
+    let controller = LockController::new(&bsq, controller_config)?;
 
-    let recipients = config
-        .recipients
-        .iter()
-        .map(
-            |recipient| match context.account_by_address(&recipient.address) {
-                Ok(account) => Ok(account.account_index()),
-                Err(_) => Err(TransactionRejectReason::InvalidAccountReference(
-                    recipient.address,
-                )),
-            },
-        )
-        .collect::<Result<Vec<_>, TransactionRejectReason>>()?;
-    let configuration =
-        LockConfiguration::new(lock_id.clone(), recipients, config.expiry, controller);
+    let recipients = match recipients {
+        CborLockRecipients::Any => LockRecipients::Any,
+        CborLockRecipients::Limited(recipients) => {
+            let recipients = recipients
+                .into_iter()
+                .map(
+                    |recipient| match context.account_by_address(&recipient.address) {
+                        Ok(account) => Ok(account.account_index()),
+                        Err(_) => Err(TransactionRejectReason::InvalidAccountReference(
+                            recipient.address,
+                        )),
+                    },
+                )
+                .collect::<Result<Vec<_>, TransactionRejectReason>>()?;
+            LockRecipients::from(recipients)
+        }
+    };
+    let configuration = LockConfiguration {
+        lock_id: lock_id.clone(),
+        recipients,
+        expiry,
+        controller,
+    };
 
     let config = get_lock_config(&bsq, &configuration).map_err(|err| {
         BlockStateFailure::Invariant(format!("Failed to get lock config for created lock: {err}"))
@@ -432,10 +448,10 @@ where
     let memo: Option<transactions::Memo> = details.memo.clone().map(transactions::Memo::from);
 
     if !lock_configuration
-        .expiry()
+        .expiry
         .is_expired(transaction_execution.timestamp())
     {
-        lock_configuration.controller().validate_operation(
+        lock_configuration.controller.validate_operation(
             &bsq,
             transaction_execution.sender_account_address(),
             transaction_execution.sender_account(),
@@ -449,14 +465,14 @@ where
             events,
             &mut token,
             account_index,
-            lock_configuration.lock_id(),
+            &lock_configuration.lock_id,
             &memo,
         )?;
         block_state.update_token(context, token)?;
     }
-    block_state.delete_lock(context, lock_configuration.lock_id())?;
+    block_state.delete_lock(context, &lock_configuration.lock_id)?;
     let event = events::LockDestroyEvent {
-        lock_id: lock_configuration.lock_id().clone(),
+        lock_id: lock_configuration.lock_id.clone(),
     };
     events.push(BlockItemEvent::LockDestroyed(event));
     Ok(())
@@ -490,7 +506,7 @@ fn remove_lock_balance_ref<C: EntityContextTypes>(
 }
 
 fn lock_configuration_keeps_alive(configuration: &LockConfiguration) -> bool {
-    match configuration.controller() {
+    match &configuration.controller {
         LockControllerConfig::SimpleV0(controller) => controller.keep_alive,
     }
 }