+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased changes
 
+## 11.2.2 (Devnet)
+
+- Add support for lock configurations with "any" specified for the `recipients` field.
+- `tokenTransfer` and `tokenBurn` operations now correctly check the available balance of the token on the account.
+
 ## 11.2.1 (Devnet)
 
 - Add support for the following "meta update" operations: